add fileClient
